requirements.txt
pandas
requests
pyarrow
duckdb
pytest
apache-airflow==2.8.4
